[refactor] 유저 상세정보 위젯화 #112
